gym-series-api-main/requirements.txt
asgiref==3.7.2
black==23.9.1
click==8.1.7
Django==4.2.5
django-stubs==4.2.4
django-stubs-ext==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
install==1.3.5
mypy==1.6.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
